+++ tools/WindowStylesReportTool/small.ico
